+++ ostap/fitting/tests/test_fitting_simfit3.py
@@ -210,6 +210,21 @@ def test_simfit3() :
         grs.draw('apl')
         graphs.append ( grs )
 
+
+    # =========================================================================
+    ## test creation of dataset
+    # =========================================================================
+    ds_gen = model_sim.generate ( nEvents = { 'N' : len ( dataset1 ) ,
+                                              'S' : len ( dataset2 ) } ,
+                                  varset  = vars  )
+
+    rg , f = model_sim.fitTo ( ds_gen , silent = True )
+    rg , f = model_sim.fitTo ( ds_gen , silent = True )
+    
+    title = 'Results of simultaneous fit to generated dataset'
+    logger.info ( '%s\n%s' % ( title , rg.table ( title = title , prefix = '# ' ) ) )
+
+ 
         
 # =============================================================================
 if '__main__' == __name__ :